requirements.txt
opencv-python-headless 
pdf2image 
pytesseract 
scikit-learn 
openai
requests
python-dotenv
#pip install paddlepaddle-gpu==3.0.0 \
#-i https://www.paddlepaddle.org.cn/packages/stable/cu118/
#pip install paddleocr



docling
pdfplumber
pillow

paddleocr
paddlepaddle-gpu